requirements.txt
numpy>=1.24.0
pandas>=2.1.0
pinecone-client>=2.2.4
openai>=1.12.0
python-dotenv>=1.0.0
python-Levenshtein>=0.23.0
tqdm>=4.66.0 